refactor: Move StepperDialog initialization directly inside the StepperDialogProvider
Let's centralize everything about StepperDialog initialization in the
same place. It will avoid complex to understand rerender in
waterfall.
